move Cvrs to testing makeContestUAfromCvrs() move to testing. move cruft out of ClcaWorkflow.
