+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/core/TestAlphaMartComparison.kt
@@ -3,7 +3,7 @@ package org.cryptobiotic.rlauxe.core
 import org.cryptobiotic.rlauxe.doublePrecision
 import org.cryptobiotic.rlauxe.sampling.ClcaNoErrorSampler
 import org.cryptobiotic.rlauxe.sampling.Sampler
-import org.cryptobiotic.rlauxe.sampling.makeContestsFromCvrs
+import org.cryptobiotic.rlauxe.util.makeContestsFromCvrs
 import org.cryptobiotic.rlauxe.sampling.makeCvrsByExactMean
 import kotlin.math.max
 import kotlin.test.Test
