+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/shangrla/TestOverstatementsFromShangrla.kt
@@ -4,7 +4,7 @@ import org.cryptobiotic.rlauxe.core.*
 import org.cryptobiotic.rlauxe.util.CvrBuilders
 import org.cryptobiotic.rlauxe.util.listToMap
 import org.cryptobiotic.rlauxe.sampling.makeCvr
-import org.cryptobiotic.rlauxe.sampling.makeFakeContest
+import org.cryptobiotic.rlauxe.util.makeFakeContest
 import org.cryptobiotic.rlauxe.util.mean2margin
 import kotlin.test.Test
 import kotlin.test.assertEquals
