+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/concur/RunRepeatedTasks.kt
@@ -24,13 +24,10 @@ import org.cryptobiotic.rlauxe.rlaplots.makeSRT
 import org.cryptobiotic.rlauxe.util.Stopwatch
 import org.cryptobiotic.rlauxe.util.mean2margin
 
-// TODO convert to use ConcurrentTaskRunner for the concurrency
-//   adapt RunTestRepeatedResult -> List<SRT>
 interface RepeatedTask {
     fun makeSampler() : Sampler
     fun makeTestFn() : RiskTestingFn
     fun makeTestParameters() : Map<String, Double>
-    // fun maxSamples() : Int
     fun name() : String
     fun N() : Int
     fun reportedMean() : Double
